server/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
boltons==23.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
colored==1.4.4
dataclasses-json==0.5.14
dnspython==2.4.2
eliot==1.14.0
eliot-tree==21.0.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==3.0.0
griffe==0.36.5
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.24.4
openai==0.28.1
orjson==3.9.9
packaging==23.2
pandas==2.0.3
pluggy==1.3.0
pydantic==1.10.9
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pyrsistent==0.19.3
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rstr==3.2.2
ruff==0.0.292
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.17
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Whoosh==2.7.4
yarl==1.9.2
zope.interface==6.1
